Add files via upload
This is the Second release for this project. v0.0.1.1
I added multi-threading for parallel downloading instead of linear downloading.
I also added a mechanism to check if file name already exists and rename the file, marking it as a copy with next available number in series, before saving the file to the folder.